Move SharedContext to ClrProfiler namespace
